+++ core/src/main/resources/reference.conf
@@ -254,8 +254,16 @@ akka.persistence.cassandra {
     # seeing events in EventsByTag queries.
     # Setting this to 0 means that tag writes will be written immediately. Batching will still happen
     # as events are buffered while a write is in progress
-    # If the tag write fails for an event the actor's persist will be failed
     flush-interval = 0ms
+
+    # Tagged events are written to a separate table after the write to the messages table has completed.
+    # If the write to the tag_views table fails it is retried. If it hasn't succeeded within this timeout
+    # then the actor will be stopped and the write will be retried again to the tag_views table when the actor
+    # is restarted
+    # A default of 4 seconds as that is greater than a typical write timeout in C* (2 seconds) and less than
+    # the default eventual consistency delay
+    # This behavior is new in 1.0.4 where before the write to the tag_views was completely asynchronous.
+    tag-write-timeout = 4s
 
     # Update the tag_scanning table with this interval. Shouldn't be done too often to
     # avoid unecessary load. The tag_scanning table keeps track of a starting point for tag
