+++ core/src/test/scala/akka/persistence/cassandra/journal/TagWritersSpec.scala
@@ -6,6 +6,7 @@ package akka.persistence.cassandra.journal
 
 import java.nio.ByteBuffer
 
+import akka.Done
 import akka.actor.{ Actor, ActorRef, ActorSystem, PoisonPill, Props }
 import akka.persistence.cassandra.Hour
 import akka.persistence.cassandra.journal.CassandraJournal.Serialized
@@ -14,14 +15,26 @@ import akka.persistence.cassandra.journal.TagWriters._
 import akka.testkit.{ ImplicitSender, TestKit, TestProbe }
 import akka.util.Timeout
 import com.datastax.oss.driver.api.core.uuid.Uuids
+import com.typesafe.config.{ Config, ConfigFactory }
 import org.scalatest.BeforeAndAfterAll
 import org.scalatest.wordspec.AnyWordSpecLike
 import org.scalatest.matchers.should.Matchers
+
 import scala.concurrent.duration._
 
-// FIXME test for replies
+object TagWritersSpec {
+  val config: Config =
+    ConfigFactory.parseString("""
+      akka {
+        use-slf4j = off
+        loglevel = INFO
+      }
+    """)
+
+}
+
 class TagWritersSpec
-    extends TestKit(ActorSystem("TagWriterSpec"))
+    extends TestKit(ActorSystem("TagWriterSpec", TagWritersSpec.config))
     with AnyWordSpecLike
     with BeforeAndAfterAll
     with ImplicitSender
@@ -54,17 +67,28 @@ class TagWritersSpec
   }
 
   "Tag writers" must {
-    "forward flush requests" in {
-      val probe = TestProbe()
-      val tagWriters = system.actorOf(testProps(defaultSettings, tag => {
-        tag shouldEqual "blue"
-        probe.ref
-      }))
 
-      tagWriters ! TagFlush("blue")
-      probe.expectMsg(Flush)
-      probe.reply(FlushComplete)
-      expectMsg(FlushComplete) // should be forwarded
+    "replies when all tag writes are complete" in {
+      val redProbe = TestProbe()
+      val blueProbe = TestProbe()
+      val probes = Map("red" -> redProbe, "blue" -> blueProbe)
+      val tagWriters = system.actorOf(testProps(defaultSettings, tag => probes(tag).ref))
+      initializePid(tagWriters)
+
+      val blueTagWrite = TagWrite("blue", List(event(pid, 0, "cat", Set("red", "blue"))))
+      val redTagWrite = TagWrite("red", List(event(pid, 0, "dog", Set("red", "blue"))))
+
+      tagWriters ! BulkTagWrite(List(blueTagWrite, redTagWrite), Nil)
+
+      expectNoMessage()
+
+      redProbe.expectMsg(redTagWrite)
+      blueProbe.expectMsg(blueTagWrite)
+
+      redProbe.reply(Done)
+      expectNoMessage()
+      blueProbe.reply(Done)
+      expectMsg(Done)
     }
 
     "flush all tag writers" in {
@@ -147,8 +171,10 @@ class TagWritersSpec
       val tagWriters = system.actorOf(testProps(defaultSettings, _ => probe.ref))
       initializePid(tagWriters)
 
-      tagWriters ! TagFlush("blue")
-      probe.expectMsg(Flush)
+      // send a tag write so that it is created
+      val preWrite = TagWrite("blue", List(dummySerialized("blue")))
+      tagWriters ! preWrite
+      probe.expectMsg(preWrite)
 
       tagWriters.tell(PassivateTagWriter("blue"), probe.ref)
       probe.expectMsg(StopTagWriter)
@@ -168,6 +194,11 @@ class TagWritersSpec
 
       probe.expectMsg(blueTagWrite)
     }
+  }
+
+  private def event(pid: String, seqNr: Long, payload: String, tags: Set[String]): Serialized = {
+    val uuid = Uuids.timeBased()
+    Serialized(pid, seqNr, ByteBuffer.wrap(payload.getBytes()), tags, "", "", 1, "", None, uuid, TimeBucket(uuid, Hour))
   }
 
   override protected def afterAll(): Unit =
